requirement.txt
flask
request
bs4
plotly
schedule
urllib